move tests to their own file, update docs
